+++ python/cudf_polars/cudf_polars/experimental/parallel.py
@@ -144,7 +144,7 @@ def get_scheduler(config_options: ConfigOptions) -> Any:
     ):  # pragma: no cover; block depends on executor type and Distributed cluster
         from distributed import get_client
 
-        from cudf_polars.experimental.dask_serialize import SerializerManager
+        from cudf_polars.experimental.dask_registers import SerializerManager
 
         client = get_client()
         SerializerManager.register_serialize()
